feat: centralized common usage @app-system/config
