+++ contracts/license/package.json
@@ -18,6 +18,7 @@
   },
   "devDependencies": {
     "@app-lib/mfe-api-kit": "*",
+    "@app-system/config": "*",
     "@rollup/plugin-replace": "^6.0.2",
     "@types/node": "^22",
     "@types/react": "^18",
