+++ lib/ui-kit/package.json
@@ -18,6 +18,7 @@
     "react-dom": "^18"
   },
   "devDependencies": {
+    "@app-system/config": "*",
     "@chromatic-com/storybook": "^3.2.6",
     "@rollup/plugin-replace": "^6.0.2",
     "@storybook/addon-essentials": "^8.6.12",
