+++ system/config/tsconfig.json
@@ -8,12 +8,15 @@
     "module": "esnext",
     "jsx": "react-jsx",
     "importHelpers": true,
-    "resolveJsonModule": true
+    "strict": true,
+    "resolveJsonModule": true,
+    "baseUrl": ".",
+    "outDir": "./dist",
+    "declaration": true,
   },
   "include": [
-    "src",
-    "../typings",
-    "./vite.config.ts"
+    "./src",
+    "../typings"
   ],
   "exclude": [
     "node_modules"
